Move tuning procedure to a module and separate generic behaviour using a parent class
